+++ client/cody-shared/src/editor/index.ts
@@ -29,4 +29,5 @@ export interface Editor {
     replaceSelection(fileName: string, selectedText: string, replacement: string): Promise<void>
     showQuickPick(labels: string[]): Promise<string | undefined>
     showWarningMessage(message: string): Promise<void>
+    showInputBox(prompt?: string): Promise<string | undefined>
 }