+++ client/cody/src/editor/vscode-editor.ts
@@ -142,4 +142,10 @@ export class VSCodeEditor implements Editor {
     public async showWarningMessage(message: string): Promise<void> {
         await vscode.window.showWarningMessage(message)
     }
+
+    public async showInputBox(prompt?: string): Promise<string | undefined> {
+        return vscode.window.showInputBox({
+            placeHolder: prompt || 'Enter here...',
+        })
+    }
 }