+++ enterprise/internal/embeddings/types.go
@@ -19,6 +19,10 @@ func (index *EmbeddingIndex) Row(n int) []int8 {
 	return index.Embeddings[n*index.ColumnDimension : (n+1)*index.ColumnDimension]
 }
 
+func (index *EmbeddingIndex) EstimateSize() int64 {
+	return int64(len(index.Embeddings) + len(index.RowMetadata)*(16+8+8) + len(index.Ranks)*4)
+}
+
 type RepoEmbeddingRowMetadata struct {
 	FileName  string `json:"fileName"`
 	StartLine int    `json:"startLine"`
@@ -32,6 +36,10 @@ type RepoEmbeddingIndex struct {
 	TextIndex EmbeddingIndex
 }
 
+func (i *RepoEmbeddingIndex) EstimateSize() int64 {
+	return i.CodeIndex.EstimateSize() + i.TextIndex.EstimateSize()
+}
+
 type ContextDetectionEmbeddingIndex struct {
 	MessagesWithAdditionalContextMeanEmbedding    []float32
 	MessagesWithoutAdditionalContextMeanEmbedding []float32
@@ -43,9 +51,9 @@ type EmbeddingSearchResults struct {
 }
 
 type EmbeddingSearchResult struct {
+	RepoName api.RepoName
+	Revision api.CommitID
 	RepoEmbeddingRowMetadata
-	// The row number in the index to correlate this result back with its source.
-	RowNum  int    `json:"rowNum"`
 	Content string `json:"content"`
 	// Experimental: Clients should not rely on any particular format of debug
 	Debug string `json:"debug,omitempty"`