+++ enterprise/internal/github_apps/store/store.go
@@ -24,6 +24,9 @@ type GitHubAppsStore interface {
 	// Update updates a GitHub App in the database and returns the updated struct.
 	Update(ctx context.Context, id int, app *types.GitHubApp) (*types.GitHubApp, error)
 
+	// Install creates a new GitHub App installation in the database.
+	Install(ctx context.Context, id, installationID int) error
+
 	// GetByID retrieves a GitHub App from the database by ID.
 	GetByID(ctx context.Context, id int) (*types.GitHubApp, error)
 
@@ -35,6 +38,9 @@ type GitHubAppsStore interface {
 
 	// WithEncryptionKey sets encryption key on store. Returns a new GitHubAppsStore
 	WithEncryptionKey(key encryption.Key) GitHubAppsStore
+
+	// List lists all GitHub Apps in the store
+	List(ctx context.Context) ([]*types.GitHubApp, error)
 }
 
 // gitHubAppStore handles storing and retrieving GitHub Apps from the database.
@@ -62,6 +68,16 @@ func (s *gitHubAppsStore) getEncryptionKey() encryption.Key {
 	return keyring.Default().GitHubAppKey
 }
 
+var scanIDAndTimes = basestore.NewFirstScanner(func(s dbutil.Scanner) (*types.GitHubApp, error) {
+	var app types.GitHubApp
+
+	err := s.Scan(
+		&app.ID,
+		&app.CreatedAt,
+		&app.UpdatedAt)
+	return &app, err
+})
+
 // Create inserts a new GitHub App into the database.
 func (s *gitHubAppsStore) Create(ctx context.Context, app *types.GitHubApp) (int, error) {
 	key := s.getEncryptionKey()
@@ -75,10 +91,10 @@ func (s *gitHubAppsStore) Create(ctx context.Context, app *types.GitHubApp) (int
 	}
 
 	query := sqlf.Sprintf(`INSERT INTO
-	    github_apps (app_id, name, slug, base_url, client_id, client_secret, private_key, encryption_key_id, logo)
-    	VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)
+	    github_apps (app_id, name, slug, base_url, app_url, client_id, client_secret, private_key, encryption_key_id, logo)
+    	VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
 		RETURNING id`,
-		app.AppID, app.Name, app.Slug, app.BaseURL, app.ClientID, clientSecret, privateKey, keyID, app.Logo)
+		app.AppID, app.Name, app.Slug, app.BaseURL, app.AppURL, app.ClientID, clientSecret, privateKey, keyID, app.Logo)
 	id, _, err := basestore.ScanFirstInt(s.Query(ctx, query))
 	return id, err
 }
@@ -89,7 +105,7 @@ func (s *gitHubAppsStore) Delete(ctx context.Context, id int) error {
 	return s.Exec(ctx, query)
 }
 
-var scanGitHubApp = basestore.NewFirstScanner(func(s dbutil.Scanner) (*types.GitHubApp, error) {
+func scanGitHubApp(s dbutil.Scanner) (*types.GitHubApp, error) {
 	var app types.GitHubApp
 
 	err := s.Scan(
@@ -98,6 +114,7 @@ var scanGitHubApp = basestore.NewFirstScanner(func(s dbutil.Scanner) (*types.Git
 		&app.Name,
 		&app.Slug,
 		&app.BaseURL,
+		&app.AppURL,
 		&app.ClientID,
 		&app.ClientSecret,
 		&app.PrivateKey,
@@ -106,24 +123,30 @@ var scanGitHubApp = basestore.NewFirstScanner(func(s dbutil.Scanner) (*types.Git
 		&app.CreatedAt,
 		&app.UpdatedAt)
 	return &app, err
-})
+}
 
-func (s *gitHubAppsStore) decrypt(ctx context.Context, app *types.GitHubApp) (*types.GitHubApp, error) {
+var (
+	scanGitHubApps     = basestore.NewSliceScanner(scanGitHubApp)
+	scanFirstGitHubApp = basestore.NewFirstScanner(scanGitHubApp)
+)
+
+func (s *gitHubAppsStore) decrypt(ctx context.Context, apps ...*types.GitHubApp) ([]*types.GitHubApp, error) {
 	key := s.getEncryptionKey()
-	var cs, pk string
 
-	cs, err := encryption.MaybeDecrypt(ctx, key, app.ClientSecret, app.EncryptionKey)
-	if err != nil {
-		return nil, err
+	for _, app := range apps {
+		cs, err := encryption.MaybeDecrypt(ctx, key, app.ClientSecret, app.EncryptionKey)
+		if err != nil {
+			return nil, err
+		}
+		app.ClientSecret = cs
+		pk, err := encryption.MaybeDecrypt(ctx, key, app.PrivateKey, app.EncryptionKey)
+		if err != nil {
+			return nil, err
+		}
+		app.PrivateKey = pk
 	}
-	app.ClientSecret = cs
-	pk, err = encryption.MaybeDecrypt(ctx, key, app.PrivateKey, app.EncryptionKey)
-	if err != nil {
-		return nil, err
-	}
-	app.PrivateKey = pk
 
-	return app, nil
+	return apps, nil
 }
 
 // Update updates a GitHub App in the database and returns the updated struct.
@@ -139,18 +162,33 @@ func (s *gitHubAppsStore) Update(ctx context.Context, id int, app *types.GitHubA
 	}
 
 	query := sqlf.Sprintf(`UPDATE github_apps
-             SET app_id = %s, name = %s, slug = %s, base_url = %s, client_id = %s, client_secret = %s, private_key = %s, encryption_key_id = %s, logo = %s, updated_at = NOW()
+             SET app_id = %s, name = %s, slug = %s, base_url = %s, app_url = %s, client_id = %s, client_secret = %s, private_key = %s, encryption_key_id = %s, logo = %s, updated_at = NOW()
              WHERE id = %s
-			 RETURNING id, app_id, name, slug, base_url, client_id, client_secret, private_key, encryption_key_id, logo, created_at, updated_at`,
-		app.AppID, app.Name, app.Slug, app.BaseURL, app.ClientID, clientSecret, privateKey, keyID, app.Logo, id)
-	app, ok, err := scanGitHubApp(s.Query(ctx, query))
+			 RETURNING id, app_id, name, slug, base_url, app_url, client_id, client_secret, private_key, encryption_key_id, logo, created_at, updated_at`,
+		app.AppID, app.Name, app.Slug, app.BaseURL, app.AppURL, app.ClientID, clientSecret, privateKey, keyID, app.Logo, id)
+	app, ok, err := scanFirstGitHubApp(s.Query(ctx, query))
 	if err != nil {
 		return nil, err
 	}
 	if !ok {
 		return nil, errors.Newf("cannot update app with id: %d because no such app exists", id)
 	}
-	return s.decrypt(ctx, app)
+	apps, err := s.decrypt(ctx, app)
+	if err != nil {
+		return nil, err
+	}
+	return apps[0], nil
+}
+
+// Install creates a new GitHub App installation in the database.
+func (s *gitHubAppsStore) Install(ctx context.Context, id, installationID int) error {
+	query := sqlf.Sprintf(`
+		INSERT INTO github_app_installs (app_id, installation_id)
+    	VALUES (%s, %s)
+		ON CONFLICT DO NOTHING
+		RETURNING id`,
+		id, installationID)
+	return s.Exec(ctx, query)
 }
 
 func (s *gitHubAppsStore) get(ctx context.Context, where *sqlf.Query) (*types.GitHubApp, error) {
@@ -160,6 +198,7 @@ func (s *gitHubAppsStore) get(ctx context.Context, where *sqlf.Query) (*types.Gi
 		name,
 		slug,
 		base_url,
+		app_url,
 		client_id,
 		client_secret,
 		private_key,
@@ -171,7 +210,7 @@ func (s *gitHubAppsStore) get(ctx context.Context, where *sqlf.Query) (*types.Gi
 	WHERE %s`
 
 	query := sqlf.Sprintf(selectQuery, where)
-	app, ok, err := scanGitHubApp(s.Query(ctx, query))
+	app, ok, err := scanFirstGitHubApp(s.Query(ctx, query))
 	if err != nil {
 		return nil, err
 	}
@@ -179,7 +218,38 @@ func (s *gitHubAppsStore) get(ctx context.Context, where *sqlf.Query) (*types.Gi
 		return nil, errors.Newf("no app exists matching criteria: %v", *where)
 	}
 
-	return s.decrypt(ctx, app)
+	apps, err := s.decrypt(ctx, app)
+	if err != nil {
+		return nil, err
+	}
+	return apps[0], nil
+}
+
+func (s *gitHubAppsStore) list(ctx context.Context, where *sqlf.Query) ([]*types.GitHubApp, error) {
+	selectQuery := `SELECT
+		id,
+		app_id,
+		name,
+		slug,
+		base_url,
+		app_url,
+		client_id,
+		client_secret,
+		private_key,
+		encryption_key_id,
+		logo,
+		created_at,
+		updated_at
+	FROM github_apps
+	WHERE %s`
+
+	query := sqlf.Sprintf(selectQuery, where)
+	apps, err := scanGitHubApps(s.Query(ctx, query))
+	if err != nil {
+		return nil, err
+	}
+
+	return s.decrypt(ctx, apps...)
 }
 
 // GetByID retrieves a GitHub App from the database by ID.
@@ -195,4 +265,9 @@ func (s *gitHubAppsStore) GetByAppID(ctx context.Context, appID int, baseURL str
 // GetBySlug retrieves a GitHub App from the database by slug and base url
 func (s *gitHubAppsStore) GetBySlug(ctx context.Context, slug string, baseURL string) (*types.GitHubApp, error) {
 	return s.get(ctx, sqlf.Sprintf(`slug = %s AND base_url = %s`, slug, baseURL))
+}
+
+// List lists all GitHub Apps in the store
+func (s *gitHubAppsStore) List(ctx context.Context) ([]*types.GitHubApp, error) {
+	return s.list(ctx, sqlf.Sprintf(`true`))
 }